pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blipshell"
version = "0.1.0"
description = "Local LLM personal assistant with persistent memory"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.4.0",
    "chromadb>=0.5.0",
    "aiosqlite>=0.20.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "rich>=13.9.0",
    "click>=8.1.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "tiktoken>=0.8.0",
    "beautifulsoup4>=4.12.0",
    "duckduckgo-search>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "ruff>=0.7.0",
]

[project.scripts]
blipshell = "blipshell.ui.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
